pyproject.toml
[project]
name = "newsletter2"
version = "0.1.0"
description = "Attempt2"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.201.1",
    "crewai-tools>=0.75.0",
    "google-search-results>=2.4.2",
    "loguru>=0.7.3",
    "openai>=1.109.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "svlearn-bootcamp>=0.1.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

# If you also build sdists and want data included:
[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    # ... other includes like README, pyproject.toml, etc.
]

[tool.hatch.build.targets.wheel]
packages = ["src/NewsLetter2"]
